api/requirements.txt
fastapi[standard]>=0.115.0,<0.116.0
sqlmodel>=0.0.27,<0.0.28
pydantic-settings>=2.0.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
cryptography>=41.0.0,<43.0.0
passlib>=1.7.4,<2.0.0
bcrypt==4.0.1
python-multipart>=0.0.6
pydantic[email]>=2.4.0,<3.0.0